Add files via upload
These two files are the supplementary file for the paper "On the admissibility of nonparametric simultaneous confidence intervals" (to be published on Symmetry, 2021). The file "larplatJune162021_github.txt" contains the R codes to run the MB and MMB methods to generate nonparametric simultaneous confidence intervals. The file "heartjune212021_github.txt" contains the R code to apply the MB and MMB methods on the heart data set to generate simultaneous confidence intervals for the regression coefficients.